Add a --check-input-hash flag
This will perform a check against already existing lockfiles determining
if the input hash of the spec has changed.  If the spec has not changed
then the lockfile is not regenerated.

This change makes some minor cleanup changed to the lockfile format for
increased clarity.  env_hash -> input_hash so its clearer what that hash
represents.