Experimental support for pip dependencies
In the absence of an external interface to pip's resolver (see e.g.
https://github.com/pypa/pip/issues/7819), this uses Poetry's
resolution logic to convert pip requirements from environment.yaml
to either transitive dependencies (in the case of env output) or
direct references (in the case of explicit output).  In explicit
mode these are emitted as comment lines that `conda-lock install`
can unpack and pass to `pip install` inside of the target environment.