Add multi-platform lockfile
conda-lock now records its solution in conda-lock.toml, in a form
roughly inspired by poetry.lock.  Each entry has a platform and a
category (e.g.  "main" or "dev"), which allows you to extract a
solution for a target platform and extras set without re-solving.
This can be done either with `conda-lock render`, creating an
environment or explicit lockfile that can be installed with vanilla
conda, or `conda-lock install` to render and install in one go.
This makes it possible for `conda-lock` to take all its configuration
from the source file.