Set encoding when manipulating stdout
When manipulating stdout please set the encoding,  as the subprocess will use the default of the system it might get another encode than unicode that python uses as a default when decoding it. That is a particular problem when using mamba because it outputs some symbols that if the default of the system is not utf-8 or other similar encodings it will raise an exception.