Fix #253: package confusion between conda and pip
Before this patch, the following would fail:
```
channels:
  - conda-forge
dependencies:
  - pip
  - dask
  - pip:
    - dask
```
The core issue seems to be that two Conda packages (`dask` and `dask-core`)
both map to the `dask` pypi package. This was causing issues later on when
assigning categories. This patch properly deals with multiple Conda packages
per pip package and addresses the naming issue.

I have verified that the above package specification works as well as the others
listed in #253.