Make the Python resolver respect any __glibc constraint
PR #541 added support for finding wheels with GLIBC 2.28. However,
as rightly pointed out in that PR:
"Note this could pose an issue if the glibc on the machine you are doing conda lock install
does not have a recent enough glibc"

This PR aims to solve this issue by propagating the GLIBC version specified
through the virtual package spec to Conda all the way down to the pypi solver.

To make all tests pass, the default GLIBC version is also raised to 2.28. We could
alternatively keep it fixed and provide a virtual-packages.yml file for the
test in PR #541 but it feels like both "defaults" should be the same (for pip and
conda).