Fix an issue when specifying full wheel URLs for a pypi package.
The pypi solver expects hashes to be provided as part of the URL but, in
most cases, if a user provides a URL for a package, they will not provide one
with a hash. This PR ignores the missing hash instead of crashing.

A test was added that fails prior to this change and succeeds after.