Drop support for Python 3.9
The current min Python version on conda-forge is 3.10. Thus
it is impossible to support 3.9 and support bugfixes. For example,
we need to upgrade ensureconda to pull in some bugfixes from
the latest version, but the latest version is no longer installable
from conda-forge. So conda-forge users have an effective min
Python version of 3.10. In order to avoid weird edge cases, it's
best that we follow suit here.